+++ history/doc.go
@@ -1,0 +1,12 @@
+// Copyright 2021 The Mellium Contributors.
+// Use of this source code is governed by the BSD 2-clause
+// license that can be found in the LICENSE file.
+
+// Package history implements fetching messages from an archive.
+package history // import "mellium.im/xmpp/history"
+
+// The namespaces used by this package, provided as a convenience.
+const (
+	NS    = `urn:xmpp:mam:2`
+	NSExt = `urn:xmpp:mam:2#extended`
+)